added all endpoint commit

+++ self_discovery/src/main/java/com/self_discovery/self_discovery/selfdiscovery/repository/InterpretationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InterpretationRepository extends JpaRepository<Interpretation, Integer> {
+public interface InterpretationRepository extends JpaRepository<Interpretation, Long> {
 }
