+++ self_discovery/src/main/java/com/self_discovery/self_discovery/selfdiscovery/repository/SectionInterpretationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SectionInterpretationRepository extends JpaRepository<SectionInterpretation, Integer> {
+public interface SectionInterpretationRepository extends JpaRepository<SectionInterpretation, Long> {
 }
